Moves prompt files into prompts folder
